Implemented custom theme creation
It is now possible, with this update, to create custom themes very
easily with some HTML/CSS knowledge.

At the current time, there is not a word of documentation, though. I
will not be able to work on PYCS for a week from now, so it'll stay that
way for a bit. However - it works!